requirements.txt
streamlit>=1.28.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
openai>=1.0.0
python-docx>=1.0.0
reportlab>=4.0.0
